refactor: move the methods to a more logical place.
